requirements.txt
streamlit
pandas
duckdb
groq
python-dotenv
watchdog